To prevent duplicate bookings, and return a valid response to the user:
When we start a booking request, we set a lock key in Redis, and put the bookign_request into the cache, with a TTL of 30 seconds.
When a booking completes, we set a response key with the booking result, and we delete the lock key
When a booking starts, we check for a lock key.  If it exists, we wait up to 30 seconds for a response key.  If the response key becomes available, we return the response.  Otherwise, we raise an error